Split validation/mod.rs into per-rule modules
Decomposes the ~4870 LOC validation/mod.rs into eleven per-rule
submodules under validation/rules/ (type_check, format, pattern, range,
length, class_types, allows_only, domain, conditional, inductive,
is_a), each owning one check_* fn plus its directly-related helpers
and its #[cfg(test)] block. Rule fns stay as Validator methods via
impl-block splitting across files (no free-function conversion); mod.rs
keeps the driver loop, ValidationError/ValidationRule, the shared
chain-walking helpers (is_instance_of_any, get_data_type_str), the
inline tests! helper functions exposed pub(in crate::validation) to
the rule test modules, and the rules without a dedicated target file
(comorphism well-formedness, merge-comorphism shape, standalone-lambda
well-typedness, universe stratification). Public surface preserved —
all existing crate::validation::* paths still resolve. Test count
unchanged at 1066 lib tests passing; clippy clean with -D warnings.
