+++ kernel/src/validation/mod.rs
@@ -16,8 +16,18 @@
 //!
 //! Validates resources in a layer against definitions reachable through
 //! the parent chain. Implements all validation rules from D1 §5.4.
+//!
+//! The per-rule check functions (Rules 3–10, 16, 17, and the
+//! class-definition / `is_a` integrity checks) live under
+//! `validation::rules` as `impl Validator` block extensions. The
+//! `Validator` struct + driver loop, the shared chain-walking helpers
+//! (`is_instance_of_any`, `is_subclass_of`, `get_data_type_str`), the
+//! universe-stratification / merge-comorphism / lambda well-typedness
+//! / comorphism well-formedness rules, and the public surface
+//! (`ValidationError`, `ValidationRule`) stay here.
 
 pub mod retroactive;
+mod rules;
 pub mod working_set;
 
 pub use retroactive::retroactive_validate;
@@ -34,48 +44,6 @@ use crate::ontology::well_known as wk;
 use std::collections::BTreeSet;
 use std::fmt;
 use std::sync::Arc;
-
-/// FormulaTerm InductiveType IRI (D32 §4). Pinned here so the
-/// formula-specific arity rule can short-circuit before doing any
-/// chain resolution work.
-const FORMULA_TERM_IRI: &str = "urn:eigenius:formulas:FormulaTerm";
-
-/// Operator.operator_arity property IRI. Convenience integer; the
-/// rank check uses it as a fast-path before the full
-/// operator_signature walk (deferred to a follow-on landing).
-const OPERATOR_ARITY_IRI: &str = "urn:eigenius:formulas:operator_arity";
-
-/// Walk the left spine of an `App(App(App(head, a₃), a₂), a₁)` tree
-/// and return `(head, [a₁, a₂, a₃])`. Spine args are emitted
-/// **right-to-left** as the spine is traversed; the caller may want
-/// to reverse if argument order matters semantically. For the arity
-/// check, only the count matters so the order is irrelevant.
-fn collect_app_spine(node: &serde_json::Value) -> (serde_json::Value, Vec<serde_json::Value>) {
-    let mut spine = Vec::new();
-    let mut cursor = node.clone();
-    loop {
-        let Some(obj) = cursor.as_object() else {
-            return (cursor, spine);
-        };
-        let ctor = obj.get("ctor").and_then(|v| v.as_str()).unwrap_or("");
-        if ctor != "App" {
-            return (cursor, spine);
-        }
-        let args = obj
-            .get("args")
-            .and_then(|v| v.as_array())
-            .cloned()
-            .unwrap_or_default();
-        if args.len() != 2 {
-            return (cursor, spine);
-        }
-        let mut iter = args.into_iter();
-        let head = iter.next().expect("len 2");
-        let arg = iter.next().expect("len 2");
-        spine.push(arg);
-        cursor = head;
-    }
-}
 
 /// A validation error describing a constraint violation.
 #[derive(Debug, Clone)]
@@ -158,7 +126,7 @@ impl fmt::Display for ValidationError {
 /// already-Arc'd layer should `Arc::clone` it; callers with a fresh
 /// `Layer` should wrap in `Arc::new`.
 pub struct Validator {
-    layer: Arc<Layer>,
+    pub(crate) layer: Arc<Layer>,
 }
 
 impl Validator {
@@ -185,23 +153,9 @@ impl Validator {
         let class_iris = resource.is_a();
         let class_refs: Vec<&Iri> = class_iris.iter().collect();
 
-        // Every resource must declare at least one class via `is_a`.
-        // Previously this was enforced incidentally by the parser
-        // rejecting empty arrays; that check was removed because empty
-        // arrays are a valid Eigon value for other properties (e.g.
-        // `urn:eigenius:query:rows` for a zero-row query result). The
-        // semantic of "an Eigon resource has at least one is_a class"
-        // is a validator rule, not a wire-format rule.
-        if class_iris.is_empty() {
-            errors.push(ValidationError {
-                resource_id: res_id.clone(),
-                property: Some(
-                    Iri::parse(crate::ontology::well_known::IS_A).expect("wk::IS_A parses"),
-                ),
-                rule: ValidationRule::MissingRequired,
-                message: "resource has no `is_a` classes".to_string(),
-            });
-        }
+        // Rule 0: every resource must declare at least one `is_a`
+        // class. (See `rules::is_a` for the full story.)
+        errors.extend(self.check_missing_is_a(resource, &res_id));
 
         // Collect effective requires/recommends from all classes + ancestors
         let (required_props, _recommended_props) = self.collect_effective_properties(&class_refs);
@@ -273,13 +227,6 @@ impl Validator {
         errors.extend(self.check_universe_stratification(resource, &res_id));
 
         // Rule 14: Class-definition reference integrity (eigenius#26).
-        // Verify that `requires` / `recommends` / `subclass_of` /
-        // `class_types` / `data_type` IRIs declared by Class and
-        // Property resources actually resolve to resources of the
-        // expected kind in the layer chain. Forward references within
-        // the same load batch are fine — by the time we run, the new
-        // layer is fully assembled and `self.layer.resolve()` walks it
-        // along with its parents.
         errors.extend(self.check_class_definition_references(resource, &res_id));
 
         // Rule 15: Comorphism well-formedness (D14 §4.5 / §5).
@@ -314,614 +261,11 @@ impl Validator {
         errors
     }
 
-    /// Collect effective `requires` and `recommends` from all classes and ancestors.
-    fn collect_effective_properties(&self, class_iris: &[&Iri]) -> (BTreeSet<Iri>, BTreeSet<Iri>) {
-        let mut required = BTreeSet::new();
-        let mut recommended = BTreeSet::new();
-
-        for class_iri in class_iris {
-            self.collect_from_class(
-                class_iri,
-                &mut required,
-                &mut recommended,
-                &mut BTreeSet::new(),
-            );
-        }
-
-        (required, recommended)
-    }
-
-    /// Recursively collect requires/recommends from a class and its ancestors.
-    fn collect_from_class(
-        &self,
-        class_iri: &Iri,
-        required: &mut BTreeSet<Iri>,
-        recommended: &mut BTreeSet<Iri>,
-        visited: &mut BTreeSet<Iri>,
-    ) {
-        if !visited.insert(class_iri.clone()) {
-            return; // Already visited (handles cycles)
-        }
-
-        if let Some(class_def) = self.layer.resolve(class_iri) {
-            // Collect requires
-            if let Some(requires_val) = class_def.get(&iri(wk::REQUIRES)) {
-                for prop_iri in requires_val.as_iri_array() {
-                    required.insert(prop_iri);
-                }
-            }
-
-            // Collect recommends
-            if let Some(recommends_val) = class_def.get(&iri(wk::RECOMMENDS)) {
-                for prop_iri in recommends_val.as_iri_array() {
-                    recommended.insert(prop_iri);
-                }
-            }
-
-            // Walk parent classes
-            if let Some(parents_val) = class_def.get(&iri(wk::PARENT_CLASSES)) {
-                for parent_iri in parents_val.as_iri_array() {
-                    self.collect_from_class(&parent_iri, required, recommended, visited);
-                }
-            }
-        }
-    }
-
-    /// Evaluate conditional_requires for all classes.
-    fn evaluate_conditional_requires(
-        &self,
-        class_iris: &[&Iri],
-        resource: &Resource,
-    ) -> (BTreeSet<Iri>, BTreeSet<Iri>) {
-        let mut required = BTreeSet::new();
-        let mut recommended = BTreeSet::new();
-
-        for class_iri in class_iris {
-            if let Some(class_def) = self.layer.resolve(class_iri) {
-                if let Some(conds) = class_def.get(&iri(wk::CONDITIONAL_REQUIRES)) {
-                    if let Some(cond_array) = conds.as_array() {
-                        for cond in cond_array {
-                            if let Value::Embedded(cond_res) = cond {
-                                self.evaluate_condition(
-                                    cond_res,
-                                    resource,
-                                    &mut required,
-                                    &mut recommended,
-                                );
-                            }
-                        }
-                    }
-                }
-            }
-        }
-
-        (required, recommended)
-    }
-
-    /// Evaluate a single ConditionalRequirement against a resource.
-    fn evaluate_condition(
-        &self,
-        condition: &Resource,
-        resource: &Resource,
-        required: &mut BTreeSet<Iri>,
-        recommended: &mut BTreeSet<Iri>,
-    ) {
-        // Get when_property — `data_type: resource`, so the canonical
-        // shape is `ResourceRef`; `as_iri` also tolerates the
-        // pre-canonical `String` shape.
-        let when_prop = match condition
-            .get(&iri(wk::WHEN_PROPERTY))
-            .and_then(|v| v.as_iri())
-        {
-            Some(i) => i,
-            None => return,
-        };
-
-        // Get has_value
-        let has_values = match condition.get(&iri(wk::HAS_VALUE)) {
-            Some(val) => val.as_iri_array(),
-            None => return,
-        };
-
-        // Check if the resource's property value matches any has_value.
-        // The value being matched is itself an IRI in either shape.
-        let resource_value = match resource.get(&when_prop) {
-            Some(v) => v,
-            None => return,
-        };
-        let matches = resource_value
-            .as_iri()
-            .map(|val_iri| has_values.contains(&val_iri))
-            .unwrap_or(false);
-
-        if matches {
-            // Apply then_requires
-            if let Some(then_req) = condition.get(&iri(wk::THEN_REQUIRES)) {
-                for prop_iri in then_req.as_iri_array() {
-                    required.insert(prop_iri);
-                }
-            }
-            // Apply then_recommends
-            if let Some(then_rec) = condition.get(&iri(wk::THEN_RECOMMENDS)) {
-                for prop_iri in then_rec.as_iri_array() {
-                    recommended.insert(prop_iri);
-                }
-            }
-        }
-    }
-
-    /// Rule 3: Type checking — value must match property's data_type.
-    fn check_type(
-        &self,
-        prop_def: &Resource,
-        value: &Value,
-        prop_iri: &Iri,
-        res_id: &Option<Iri>,
-    ) -> Vec<ValidationError> {
-        let dt = match self.get_data_type_str(prop_def) {
-            Some(dt) => dt,
-            None => return vec![], // No data_type defined, skip
-        };
-
-        let ok = match dt.as_str() {
-            wk::STRING => matches!(value, Value::String(_)),
-            wk::INTEGER => matches!(value, Value::Integer(_)),
-            wk::FLOAT => matches!(value, Value::Float(_) | Value::Integer(_)),
-            wk::BOOLEAN => matches!(value, Value::Boolean(_)),
-            wk::RESOURCE => {
-                // Post-canonicalisation (see `LayerBuilder::build` ->
-                // `canonicalise_resource_refs`), every resource-typed
-                // value on a committed layer is either `ResourceRef`
-                // (an IRI) or `Embedded` (an inlined Resource).
-                // `Value::String` for a `data_type: resource`
-                // property is a malformed declaration the
-                // canonicaliser couldn't normalise (typically because
-                // the property def is missing or its data_type isn't
-                // resolvable) — flag it as a type mismatch rather
-                // than silently accepting a non-canonical shape.
-                //
-                // When `class_types` declares an `InductiveType`, also
-                // accept `Value::Json` — the tagged-dict carrier for
-                // inductive values. The deeper structural check
-                // (ctor / arg_types) runs in `check_class_types`,
-                // mirroring the `core:inductive` split.
-                if self.class_types_inductive_target(prop_def).is_some() {
-                    matches!(
-                        value,
-                        Value::ResourceRef(_) | Value::Embedded(_) | Value::Json(_)
-                    )
-                } else {
-                    matches!(value, Value::ResourceRef(_) | Value::Embedded(_))
-                }
-            }
-            wk::RESOURCE_ARRAY => match value {
-                Value::Array(arr) => {
-                    if self.class_types_inductive_target(prop_def).is_some() {
-                        arr.iter().all(|v| {
-                            matches!(
-                                v,
-                                Value::ResourceRef(_) | Value::Embedded(_) | Value::Json(_)
-                            )
-                        })
-                    } else {
-                        arr.iter()
-                            .all(|v| matches!(v, Value::ResourceRef(_) | Value::Embedded(_)))
-                    }
-                }
-                _ => false,
-            },
-            wk::VALUE_ARRAY => matches!(value, Value::Array(_)),
-            wk::JSON => true, // Any value is valid for JSON
-            wk::INDUCTIVE => {
-                // Wire-level shape check: an inductive value lands as
-                // either a `Value::Json` carrying the tagged-dict tree
-                // or a `Value::Embedded` resource. The deeper
-                // structural type-check (ctor exists on declared
-                // InductiveType, arg shapes match `arg_types`) lives in
-                // `check_inductive_value` (rule 16) — same split as
-                // `check_class_types` for `core:resource`.
-                matches!(value, Value::Json(_) | Value::Embedded(_))
-            }
-            _ => true, // Unknown data type, skip
-        };
-
-        if ok {
-            vec![]
-        } else {
-            vec![ValidationError {
-                resource_id: res_id.clone(),
-                property: Some(prop_iri.clone()),
-                rule: ValidationRule::TypeMismatch,
-                message: format!("expected data_type '{dt}', got incompatible value"),
-            }]
-        }
-    }
-
-    /// Rule 4: Format checking.
-    fn check_format(
-        &self,
-        prop_def: &Resource,
-        value: &Value,
-        prop_iri: &Iri,
-        res_id: &Option<Iri>,
-    ) -> Vec<ValidationError> {
-        let format_str = match prop_def.get(&iri(wk::FORMAT_PROP)) {
-            Some(Value::String(s)) => s.as_str(),
-            _ => return vec![],
-        };
-
-        let string_val = match value.as_str() {
-            Some(s) => s,
-            None => return vec![], // Not a string, type checking handles this
-        };
-
-        let valid = match format_str {
-            wk::FMT_DATE => is_valid_date(string_val),
-            wk::FMT_DATETIME => is_valid_datetime(string_val),
-            wk::FMT_TIME => is_valid_time(string_val),
-            wk::FMT_IRI => Iri::parse(string_val).is_ok(),
-            wk::FMT_UUID => is_valid_uuid(string_val),
-            wk::FMT_REGEX => regex::Regex::new(string_val).is_ok(),
-            _ => true, // Unknown format, skip
-        };
-
-        if valid {
-            vec![]
-        } else {
-            vec![ValidationError {
-                resource_id: res_id.clone(),
-                property: Some(prop_iri.clone()),
-                rule: ValidationRule::FormatViolation,
-                message: format!("value '{string_val}' does not match format '{format_str}'"),
-            }]
-        }
-    }
-
-    /// Rule 5: Pattern checking.
-    fn check_pattern(
-        &self,
-        prop_def: &Resource,
-        value: &Value,
-        prop_iri: &Iri,
-        res_id: &Option<Iri>,
-    ) -> Vec<ValidationError> {
-        let pattern_str = match prop_def.get(&iri(wk::PATTERN)) {
-            Some(Value::String(s)) => s.as_str(),
-            _ => return vec![],
-        };
-
-        let string_val = match value.as_str() {
-            Some(s) => s,
-            None => return vec![],
-        };
-
-        // Full match: wrap in ^...$
-        let full_pattern = format!("^(?:{pattern_str})$");
-        match regex::Regex::new(&full_pattern) {
-            Ok(re) => {
-                if re.is_match(string_val) {
-                    vec![]
-                } else {
-                    vec![ValidationError {
-                        resource_id: res_id.clone(),
-                        property: Some(prop_iri.clone()),
-                        rule: ValidationRule::PatternViolation,
-                        message: format!(
-                            "value '{string_val}' does not match pattern '{pattern_str}'"
-                        ),
-                    }]
-                }
-            }
-            Err(_) => vec![], // Invalid regex in property def — skip (should be caught by format validation on the property itself)
-        }
-    }
-
-    /// Rule 6: Range checking (min_value/max_value).
-    fn check_range(
-        &self,
-        prop_def: &Resource,
-        value: &Value,
-        prop_iri: &Iri,
-        res_id: &Option<Iri>,
-    ) -> Vec<ValidationError> {
-        let num_val = match value {
-            Value::Integer(n) => *n as f64,
-            Value::Float(f) => *f,
-            _ => return vec![],
-        };
-
-        let mut errors = Vec::new();
-
-        if let Some(Value::Float(min)) = prop_def.get(&iri(wk::MIN_VALUE)) {
-            if num_val < *min {
-                errors.push(ValidationError {
-                    resource_id: res_id.clone(),
-                    property: Some(prop_iri.clone()),
-                    rule: ValidationRule::RangeViolation,
-                    message: format!("value {num_val} is less than minimum {min}"),
-                });
-            }
-        }
-        if let Some(Value::Integer(min)) = prop_def.get(&iri(wk::MIN_VALUE)) {
-            if num_val < *min as f64 {
-                errors.push(ValidationError {
-                    resource_id: res_id.clone(),
-                    property: Some(prop_iri.clone()),
-                    rule: ValidationRule::RangeViolation,
-                    message: format!("value {num_val} is less than minimum {min}"),
-                });
-            }
-        }
-
-        if let Some(Value::Float(max)) = prop_def.get(&iri(wk::MAX_VALUE)) {
-            if num_val > *max {
-                errors.push(ValidationError {
-                    resource_id: res_id.clone(),
-                    property: Some(prop_iri.clone()),
-                    rule: ValidationRule::RangeViolation,
-                    message: format!("value {num_val} is greater than maximum {max}"),
-                });
-            }
-        }
-        if let Some(Value::Integer(max)) = prop_def.get(&iri(wk::MAX_VALUE)) {
-            if num_val > *max as f64 {
-                errors.push(ValidationError {
-                    resource_id: res_id.clone(),
-                    property: Some(prop_iri.clone()),
-                    rule: ValidationRule::RangeViolation,
-                    message: format!("value {num_val} is greater than maximum {max}"),
-                });
-            }
-        }
-
-        errors
-    }
-
-    /// Rule 7: Length checking (min_length/max_length).
-    fn check_length(
-        &self,
-        prop_def: &Resource,
-        value: &Value,
-        prop_iri: &Iri,
-        res_id: &Option<Iri>,
-    ) -> Vec<ValidationError> {
-        let len = match value {
-            Value::String(s) => s.chars().count(),
-            Value::Array(arr) => arr.len(),
-            _ => return vec![],
-        };
-
-        let mut errors = Vec::new();
-
-        if let Some(Value::Integer(min)) = prop_def.get(&iri(wk::MIN_LENGTH)) {
-            if (len as i64) < *min {
-                errors.push(ValidationError {
-                    resource_id: res_id.clone(),
-                    property: Some(prop_iri.clone()),
-                    rule: ValidationRule::LengthViolation,
-                    message: format!("length {len} is less than minimum {min}"),
-                });
-            }
-        }
-
-        if let Some(Value::Integer(max)) = prop_def.get(&iri(wk::MAX_LENGTH)) {
-            if (len as i64) > *max {
-                errors.push(ValidationError {
-                    resource_id: res_id.clone(),
-                    property: Some(prop_iri.clone()),
-                    rule: ValidationRule::LengthViolation,
-                    message: format!("length {len} is greater than maximum {max}"),
-                });
-            }
-        }
-
-        errors
-    }
-
-    /// Rule 8: Class type checking.
-    ///
-    /// `class_types` may name either a `Class` (the historical case —
-    /// the value must be a ResourceRef/Embedded whose `is_a` matches)
-    /// or an `InductiveType` (Option A unification — the value is a
-    /// tagged-dict tree carried by `Value::Json`, and we dispatch to
-    /// the inductive walker). Per the singleton constraint that
-    /// already applies to `data_type: core:inductive`, an
-    /// InductiveType `class_types` must be the sole entry; mixed
-    /// Class/InductiveType lists are not a defined shape.
-    fn check_class_types(
-        &self,
-        prop_def: &Resource,
-        value: &Value,
-        prop_iri: &Iri,
-        res_id: &Option<Iri>,
-    ) -> Vec<ValidationError> {
-        let allowed_classes = match prop_def.get(&iri(wk::CLASS_TYPES)) {
-            Some(val) => val.as_iri_array(),
-            None => return vec![],
-        };
-
-        if allowed_classes.is_empty() {
-            return vec![];
-        }
-
-        // InductiveType branch: walk the tagged-dict tree(s).
-        // Skipping non-Json elements here is intentional — wire-shape
-        // (`check_type`) handles whether a Ref/Embedded is admissible
-        // for the property's data_type; deep class-membership of
-        // stored inductive instances is not a v1 use case.
-        //
-        // For `data_type: core:inductive`, `check_inductive_value`
-        // (rule 16) owns the walk + the singleton precondition; we
-        // defer to it to avoid duplicate diagnostics.
-        let dt_is_inductive = self
-            .get_data_type_str(prop_def)
-            .map(|dt| dt == wk::INDUCTIVE)
-            .unwrap_or(false);
-        if !dt_is_inductive {
-            if let Some(inductive_type) = self.class_types_inductive_target(prop_def) {
-                let mut errors = Vec::new();
-                match value {
-                    Value::Json(_) => {
-                        self.walk_inductive_value(
-                            value,
-                            &inductive_type,
-                            prop_iri.as_str().to_string(),
-                            res_id,
-                            &mut errors,
-                        );
-                    }
-                    Value::Array(arr) => {
-                        for (i, v) in arr.iter().enumerate() {
-                            if matches!(v, Value::Json(_)) {
-                                let path = format!("{prop_iri}[{i}]");
-                                self.walk_inductive_value(
-                                    v,
-                                    &inductive_type,
-                                    path,
-                                    res_id,
-                                    &mut errors,
-                                );
-                            }
-                        }
-                    }
-                    _ => {}
-                }
-                return errors;
-            }
-        }
-
-        let allowed_refs: Vec<&Iri> = allowed_classes.iter().collect();
-
-        let mut errors = Vec::new();
-        let values_to_check = match value {
-            Value::String(_) | Value::ResourceRef(_) | Value::Embedded(_) => vec![value],
-            Value::Array(arr) => arr.iter().collect(),
-            _ => return vec![],
-        };
-
-        for v in values_to_check {
-            // Embedded resources are checked directly against the
-            // allowed-class set; IRI references (in either canonical
-            // `ResourceRef` or pre-canonical `String` shape) are
-            // resolved through the chain first.
-            if let Value::Embedded(embedded) = v {
-                if !self.is_instance_of_any(embedded, &allowed_refs) {
-                    let actual = format_is_a_list(embedded.is_a());
-                    let allowed = format_iri_refs(&allowed_refs);
-                    errors.push(ValidationError {
-                        resource_id: res_id.clone(),
-                        property: Some(prop_iri.clone()),
-                        rule: ValidationRule::ClassTypeMismatch,
-                        message: format!(
-                            "embedded value on property '{prop_iri}' has is_a {actual} but must be an instance of one of: {allowed}"
-                        ),
-                    });
-                }
-                continue;
-            }
-            if let Some(ref_iri) = v.as_iri() {
-                if let Some(referenced) = self.layer.resolve(&ref_iri) {
-                    if !self.is_instance_of_any(&referenced, &allowed_refs) {
-                        let actual = format_is_a_list(referenced.is_a());
-                        let allowed = format_iri_refs(&allowed_refs);
-                        errors.push(ValidationError {
-                            resource_id: res_id.clone(),
-                            property: Some(prop_iri.clone()),
-                            rule: ValidationRule::ClassTypeMismatch,
-                            message: format!(
-                                "referenced resource '{ref_iri}' has is_a {actual} but must be an instance of one of: {allowed}"
-                            ),
-                        });
-                    }
-                }
-                // If we can't resolve, skip — might be external.
-            }
-        }
-
-        errors
-    }
-
-    /// Rule 9: Allowed values checking.
-    fn check_allows_only(
-        &self,
-        prop_def: &Resource,
-        value: &Value,
-        prop_iri: &Iri,
-        res_id: &Option<Iri>,
-    ) -> Vec<ValidationError> {
-        let allowed = match prop_def.get(&iri(wk::ALLOWS_ONLY)) {
-            Some(val) => val.as_iri_array(),
-            None => return vec![],
-        };
-
-        if allowed.is_empty() {
-            return vec![];
-        }
-
-        let allowed_set: BTreeSet<Iri> = allowed.into_iter().collect();
-        let mut errors = Vec::new();
-
-        // Collect candidate IRIs to test against the allows_only set.
-        // Single-value properties hold one IRI directly; resource_array
-        // properties hold a `Value::Array` of IRI elements. `as_iri`
-        // accepts both canonical `ResourceRef` and pre-canonical
-        // `String` shapes.
-        let refs_to_check: Vec<Iri> = match value {
-            Value::Array(arr) => arr.iter().filter_map(|v| v.as_iri()).collect(),
-            single => single.as_iri().map(|i| vec![i]).unwrap_or_default(),
-        };
-
-        for ref_iri in refs_to_check {
-            {
-                if !allowed_set.contains(&ref_iri) {
-                    errors.push(ValidationError {
-                        resource_id: res_id.clone(),
-                        property: Some(prop_iri.clone()),
-                        rule: ValidationRule::AllowedValueViolation,
-                        message: format!("value '{ref_iri}' is not in the allows_only set"),
-                    });
-                }
-            }
-        }
-
-        errors
-    }
-
-    /// Rule 10: Domain checking.
-    fn check_domain(
-        &self,
-        prop_def: &Resource,
-        resource: &Resource,
-        prop_iri: &Iri,
-        res_id: &Option<Iri>,
-    ) -> Vec<ValidationError> {
-        let domain_classes = match prop_def.get(&iri(wk::DOMAIN)) {
-            Some(val) => val.as_iri_array(),
-            None => return vec![], // No domain constraint
-        };
-
-        if domain_classes.is_empty() {
-            return vec![];
-        }
-
-        let domain_refs: Vec<&Iri> = domain_classes.iter().collect();
-        if self.is_instance_of_any(resource, &domain_refs) {
-            vec![]
-        } else {
-            vec![ValidationError {
-                resource_id: res_id.clone(),
-                property: Some(prop_iri.clone()),
-                rule: ValidationRule::DomainViolation,
-                message: format!(
-                    "property '{prop_iri}' is not allowed on this resource type (domain restriction)"
-                ),
-            }]
-        }
-    }
+    // ── Shared chain-walking helpers used across multiple rule files ──
 
     /// Check if a resource is an instance of any of the given classes,
     /// considering subclass relationships.
-    fn is_instance_of_any(&self, resource: &Resource, classes: &[&Iri]) -> bool {
+    pub(crate) fn is_instance_of_any(&self, resource: &Resource, classes: &[&Iri]) -> bool {
         let resource_classes = resource.is_a();
         for res_class in &resource_classes {
             for allowed_class in classes {
@@ -964,7 +308,7 @@ impl Validator {
     }
 
     /// Extract the data_type IRI string from a property definition.
-    fn get_data_type_str(&self, prop_def: &Resource) -> Option<String> {
+    pub(crate) fn get_data_type_str(&self, prop_def: &Resource) -> Option<String> {
         // `data_type` is a `data_type: resource` property, so after
         // `LayerBuilder::canonicalise_resource_refs` runs the value is
         // a `Value::ResourceRef`. Accept the (legacy) `Value::String`
@@ -974,540 +318,6 @@ impl Validator {
             .get(&iri(wk::DATA_TYPE_PROP))
             .and_then(|v| v.as_iri())
             .map(|i| i.as_str().to_string())
-    }
-
-    /// Resolve `class_types` to an `InductiveType` resource when the
-    /// property declares exactly one entry pointing to one. Returns
-    /// `None` for the Class case (the original `class_types`
-    /// semantics) and for mixed/empty lists. Powers the Option A
-    /// unification across `core:resource`, `core:resource_array`,
-    /// and (implicitly, via the singleton constraint) `core:inductive`.
-    fn class_types_inductive_target(&self, prop_def: &Resource) -> Option<Arc<Resource>> {
-        let class_iris = prop_def.get(&iri(wk::CLASS_TYPES))?.as_iri_array();
-        if class_iris.len() != 1 {
-            return None;
-        }
-        let target = self.layer.resolve(&class_iris[0])?;
-        if target.is_instance_of(&iri(wk::INDUCTIVE_TYPE)) {
-            Some(target)
-        } else {
-            None
-        }
-    }
-
-    /// Rule 14: Class-definition reference integrity (eigenius#26).
-    ///
-    /// Rule 16: Inductive value type-checking (D32 §3.5).
-    ///
-    /// When a property has `data_type: core:inductive`, its `class_types`
-    /// must declare exactly one `core:InductiveType`, and the value must
-    /// be a tagged-dict tree (`{ "ctor": ..., "args": [...] }`) whose
-    /// every node corresponds to a ctor declared on the inductive and
-    /// whose every arg matches the ctor's declared `arg_types[i].type_name`.
-    /// Errors carry structured paths so users see
-    /// `term.args[0].args[1]: ctor 'foo' not declared on FormulaTerm`.
-    fn check_inductive_value(
-        &self,
-        prop_def: &Resource,
-        value: &Value,
-        prop_iri: &Iri,
-        res_id: &Option<Iri>,
-    ) -> Vec<ValidationError> {
-        let dt = match self.get_data_type_str(prop_def) {
-            Some(dt) => dt,
-            None => return vec![],
-        };
-        if dt != wk::INDUCTIVE {
-            return vec![];
-        }
-
-        let allowed = match prop_def.get(&iri(wk::CLASS_TYPES)) {
-            Some(val) => val.as_iri_array(),
-            None => Vec::new(),
-        };
-        if allowed.len() != 1 {
-            return vec![ValidationError {
-                resource_id: res_id.clone(),
-                property: Some(prop_iri.clone()),
-                rule: ValidationRule::TypeMismatch,
-                message: format!(
-                    "data_type 'core:inductive' requires exactly one `class_types` entry naming an InductiveType (got {})",
-                    allowed.len()
-                ),
-            }];
-        }
-        let ind_iri = allowed.into_iter().next().expect("len 1");
-
-        let ind_type = match self.layer.resolve(&ind_iri) {
-            Some(r) => r,
-            None => {
-                return vec![ValidationError {
-                    resource_id: res_id.clone(),
-                    property: Some(prop_iri.clone()),
-                    rule: ValidationRule::UnresolvedClassReference,
-                    message: format!(
-                        "inductive type '{ind_iri}' on `class_types` not found in chain"
-                    ),
-                }];
-            }
-        };
-        if !ind_type.is_instance_of(&iri(wk::INDUCTIVE_TYPE)) {
-            return vec![ValidationError {
-                resource_id: res_id.clone(),
-                property: Some(prop_iri.clone()),
-                rule: ValidationRule::TypeMismatch,
-                message: format!(
-                    "`class_types` IRI '{ind_iri}' is not an InductiveType — `data_type: core:inductive` requires one"
-                ),
-            }];
-        }
-
-        let mut errors = Vec::new();
-        self.walk_inductive_value(
-            value,
-            &ind_type,
-            prop_iri.as_str().to_string(),
-            res_id,
-            &mut errors,
-        );
-        errors
-    }
-
-    /// Recursively type-check an inductive value tree against an
-    /// `InductiveType` resource. `path` accumulates a structured trace
-    /// (`term.args[0].args[1]`) for diagnostic clarity.
-    fn walk_inductive_value(
-        &self,
-        value: &Value,
-        inductive_type: &Resource,
-        path: String,
-        res_id: &Option<Iri>,
-        out: &mut Vec<ValidationError>,
-    ) {
-        let json = match value {
-            Value::Json(j) => j,
-            other => {
-                out.push(ValidationError {
-                    resource_id: res_id.clone(),
-                    property: None,
-                    rule: ValidationRule::InductiveValueMismatch,
-                    message: format!(
-                        "{path}: expected JSON tagged-dict for inductive value, got {other:?}"
-                    ),
-                });
-                return;
-            }
-        };
-
-        let obj = match json.as_object() {
-            Some(o) => o,
-            None => {
-                out.push(ValidationError {
-                    resource_id: res_id.clone(),
-                    property: None,
-                    rule: ValidationRule::InductiveValueMismatch,
-                    message: format!("{path}: inductive value must be a JSON object"),
-                });
-                return;
-            }
-        };
-
-        let ctor_name = match obj.get("ctor").and_then(serde_json::Value::as_str) {
-            Some(s) => s,
-            None => {
-                out.push(ValidationError {
-                    resource_id: res_id.clone(),
-                    property: None,
-                    rule: ValidationRule::InductiveValueMismatch,
-                    message: format!("{path}: inductive value missing string `ctor` field"),
-                });
-                return;
-            }
-        };
-
-        let args_array = obj
-            .get("args")
-            .and_then(serde_json::Value::as_array)
-            .cloned()
-            .unwrap_or_default();
-
-        // Find the ctor declaration on the inductive type.
-        let ctors_value = match inductive_type.get(&iri(wk::CTORS)) {
-            Some(v) => v,
-            None => {
-                out.push(ValidationError {
-                    resource_id: res_id.clone(),
-                    property: None,
-                    rule: ValidationRule::InductiveValueMismatch,
-                    message: format!(
-                        "{path}: InductiveType `{}` has no `ctors` declared",
-                        inductive_type.id().map(|i| i.as_str()).unwrap_or("?"),
-                    ),
-                });
-                return;
-            }
-        };
-        let ctor_arr = match ctors_value {
-            Value::Array(a) => a,
-            _ => return, // Earlier rules will have flagged this
-        };
-
-        let matching_ctor = ctor_arr.iter().find_map(|c| match c {
-            Value::Embedded(r) => {
-                let name = r
-                    .get(&iri(wk::CTOR_NAME))
-                    .and_then(Value::as_str)
-                    .unwrap_or("");
-                if name == ctor_name {
-                    Some(r.as_ref())
-                } else {
-                    None
-                }
-            }
-            _ => None,
-        });
-
-        let ctor = match matching_ctor {
-            Some(c) => c,
-            None => {
-                out.push(ValidationError {
-                    resource_id: res_id.clone(),
-                    property: None,
-                    rule: ValidationRule::InductiveValueMismatch,
-                    message: format!(
-                        "{path}: ctor `{ctor_name}` not declared on InductiveType `{}`",
-                        inductive_type.id().map(|i| i.as_str()).unwrap_or("?"),
-                    ),
-                });
-                return;
-            }
-        };
-
-        let arg_types: Vec<Resource> = match ctor.get(&iri(wk::ARG_TYPES)) {
-            Some(Value::Array(a)) => a
-                .iter()
-                .filter_map(|v| match v {
-                    Value::Embedded(r) => Some(r.as_ref().clone()),
-                    _ => None,
-                })
-                .collect(),
-            _ => Vec::new(),
-        };
-
-        if args_array.len() != arg_types.len() {
-            out.push(ValidationError {
-                resource_id: res_id.clone(),
-                property: None,
-                rule: ValidationRule::InductiveValueMismatch,
-                message: format!(
-                    "{path}: ctor `{ctor_name}` expects {} arg(s), got {}",
-                    arg_types.len(),
-                    args_array.len(),
-                ),
-            });
-            return;
-        }
-
-        for (i, (arg_value, arg_type_decl)) in args_array.iter().zip(arg_types.iter()).enumerate() {
-            let type_name = arg_type_decl
-                .get(&iri(wk::TYPE_NAME))
-                .and_then(Value::as_str)
-                .unwrap_or("");
-            let child_path = format!("{path}.args[{i}]");
-            self.check_inductive_arg(arg_value, type_name, child_path, res_id, out);
-        }
-    }
-
-    /// Validate one argument value against the `type_name` declared on
-    /// its `InductiveArgType`. Dispatches on whether `type_name` resolves
-    /// to a primitive `DataType`, a `Class`, an `InductiveType`, or a
-    /// bare type-parameter name (deferred — parameter-aware checking
-    /// lands when parametric inductives have their first chain
-    /// consumer; v1 callers use only monomorphic inductives).
-    fn check_inductive_arg(
-        &self,
-        arg_value: &serde_json::Value,
-        type_name: &str,
-        path: String,
-        res_id: &Option<Iri>,
-        out: &mut Vec<ValidationError>,
-    ) {
-        // Try to parse as IRI and resolve. If it doesn't parse or
-        // doesn't resolve, treat as an unbound parameter name and
-        // skip (v1 deferral).
-        let type_iri = match Iri::parse(type_name) {
-            Ok(i) => i,
-            Err(_) => return, // Bare parameter name; deferred per v1.
-        };
-
-        // Primitive type IRIs are well-known; check inline.
-        let ok = match type_name {
-            wk::STRING => arg_value.is_string(),
-            wk::INTEGER => arg_value.is_i64(),
-            wk::FLOAT => arg_value.is_number(),
-            wk::BOOLEAN => arg_value.is_boolean(),
-            _ => {
-                // Resolve to a chain Resource.
-                let referent = match self.layer.resolve(&type_iri) {
-                    Some(r) => r,
-                    None => return, // Treat as unbound parameter; deferred.
-                };
-                // InductiveType: recurse.
-                if referent.is_instance_of(&iri(wk::INDUCTIVE_TYPE)) {
-                    self.walk_inductive_value(
-                        &Value::Json(arg_value.clone()),
-                        &referent,
-                        path,
-                        res_id,
-                        out,
-                    );
-                    return;
-                }
-                // Class: arg is an embedded resource ref or IRI string —
-                // structural shape only; deeper class-type checking
-                // would duplicate `check_class_types` and is deferred.
-                // For v1, accept any string (IRI ref) or object (embedded).
-                if referent.is_instance_of(&iri(wk::CLASS)) {
-                    arg_value.is_string() || arg_value.is_object()
-                } else {
-                    // Unknown referent kind — skip silently.
-                    true
-                }
-            }
-        };
-
-        if !ok {
-            out.push(ValidationError {
-                resource_id: res_id.clone(),
-                property: None,
-                rule: ValidationRule::InductiveValueMismatch,
-                message: format!("{path}: value does not match declared `type_name` `{type_name}`"),
-            });
-        }
-    }
-
-    /// Rule 17: FormulaTerm App-spine arity check (D32 §5.4).
-    ///
-    /// When a property's value is a FormulaTerm whose outer ctor is
-    /// `App`, walk the left spine to find the head. If the head is an
-    /// `OpRef(iri)` whose target resolves to an `Operator` resource
-    /// with a declared `operator_arity`, confirm the App spine
-    /// supplies exactly that many arguments. This catches typos like
-    /// `App(OpRef("add"), x)` (one arg short) at commit time rather
-    /// than at dispatch.
-    ///
-    /// Type-of-each-arg checking against the operator's full
-    /// `operator_signature` (a Pi chain over FormulaTerm) is a
-    /// follow-on landing — v1 ships arity-only.
-    fn check_formula_term_arity(
-        &self,
-        prop_def: &Resource,
-        value: &Value,
-        prop_iri: &Iri,
-        res_id: &Option<Iri>,
-    ) -> Vec<ValidationError> {
-        // Only fire on properties whose value is a FormulaTerm.
-        let dt = match self.get_data_type_str(prop_def) {
-            Some(dt) => dt,
-            None => return vec![],
-        };
-        if dt != wk::INDUCTIVE {
-            return vec![];
-        }
-        let allowed = match prop_def.get(&iri(wk::CLASS_TYPES)) {
-            Some(val) => val.as_iri_array(),
-            None => return vec![],
-        };
-        if allowed.len() != 1 {
-            return vec![];
-        }
-        if allowed[0].as_str() != FORMULA_TERM_IRI {
-            return vec![];
-        }
-
-        let json = match value {
-            Value::Json(j) => j,
-            _ => return vec![],
-        };
-
-        let mut errors = Vec::new();
-        self.walk_formula_term_app_arity(json, prop_iri.as_str().to_string(), res_id, &mut errors);
-        errors
-    }
-
-    /// Walk a FormulaTerm value tree. When entering an `App` node,
-    /// resolve the *whole* left spine to its head + spine args and
-    /// arity-check once. Then recurse only into the spine args
-    /// (each of which may itself be a sub-tree carrying nested
-    /// applications). The intermediate `App` nodes inside the spine
-    /// are *not* re-checked — they're partial applications, not
-    /// complete operator invocations.
-    ///
-    /// For non-App nodes, recurse into every arg so nested
-    /// applications buried in `Lam(_, ty, body)` etc. still get
-    /// checked.
-    fn walk_formula_term_app_arity(
-        &self,
-        node: &serde_json::Value,
-        path: String,
-        res_id: &Option<Iri>,
-        out: &mut Vec<ValidationError>,
-    ) {
-        let obj = match node.as_object() {
-            Some(o) => o,
-            None => return,
-        };
-        let ctor = obj
-            .get("ctor")
-            .and_then(serde_json::Value::as_str)
-            .unwrap_or("");
-
-        if ctor == "App" {
-            let (head, spine_args) = collect_app_spine(node);
-            if let Some(head_obj) = head.as_object() {
-                let head_ctor = head_obj
-                    .get("ctor")
-                    .and_then(serde_json::Value::as_str)
-                    .unwrap_or("");
-                if head_ctor == "OpRef" {
-                    let op_iri_s = head_obj
-                        .get("args")
-                        .and_then(serde_json::Value::as_array)
-                        .and_then(|a| a.first())
-                        .and_then(serde_json::Value::as_str);
-                    if let Some(op_iri_s) = op_iri_s {
-                        if let Ok(op_iri) = Iri::parse(op_iri_s) {
-                            if let Some(op_resource) = self.layer.resolve(&op_iri) {
-                                if let Some(arity_value) = op_resource.get(&iri(OPERATOR_ARITY_IRI))
-                                {
-                                    if let Some(arity) = arity_value.as_integer() {
-                                        if (arity as usize) != spine_args.len() {
-                                            out.push(ValidationError {
-                                                resource_id: res_id.clone(),
-                                                property: None,
-                                                rule: ValidationRule::OperatorArityMismatch,
-                                                message: format!(
-                                                    "{path}: operator `{op_iri_s}` declares arity {arity}; App spine supplies {} arg(s)",
-                                                    spine_args.len(),
-                                                ),
-                                            });
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-            // Recurse only into spine args (NOT into intermediate
-            // App heads — those are partial applications counted by
-            // the spine, not separate invocations to arity-check).
-            // `collect_app_spine` returns args right-to-left from
-            // the deepest App; reverse so paths read left-to-right.
-            let spine_left_to_right: Vec<&serde_json::Value> = spine_args.iter().rev().collect();
-            for (i, arg) in spine_left_to_right.iter().enumerate() {
-                let child_path = format!("{path}.args[{i}]");
-                self.walk_formula_term_app_arity(arg, child_path, res_id, out);
-            }
-            return;
-        }
-
-        // Non-App node: recurse into every arg so nested applications
-        // inside Lam/Pi bodies, OpRef IRIs (no recursion needed —
-        // they're string args), etc. get checked too.
-        let args = obj
-            .get("args")
-            .and_then(serde_json::Value::as_array)
-            .cloned()
-            .unwrap_or_default();
-        for (i, arg) in args.iter().enumerate() {
-            let child_path = format!("{path}.args[{i}]");
-            self.walk_formula_term_app_arity(arg, child_path, res_id, out);
-        }
-    }
-
-    /// For Class resources, every IRI in `requires` / `recommends` must
-    /// resolve to a `core:Property` and every IRI in `subclass_of` must
-    /// resolve to a `core:Class`. For Property resources, every IRI in
-    /// `class_types` must resolve to a `core:Class` and `data_type` (if
-    /// present) must resolve to a `core:DataType`.
-    ///
-    /// Without this, a typo in `requires patent:innovation_category`
-    /// (vs `invention_category`) commits cleanly and only fails much
-    /// later at instance validation or program execution time, far
-    /// from the offending declaration.
-    fn check_class_definition_references(
-        &self,
-        resource: &Resource,
-        res_id: &Option<Iri>,
-    ) -> Vec<ValidationError> {
-        let mut errors = Vec::new();
-        let is_class = resource.is_instance_of(&iri(wk::CLASS));
-        let is_property = resource.is_instance_of(&iri(wk::PROPERTY));
-        if !is_class && !is_property {
-            return errors;
-        }
-
-        if is_class {
-            self.check_array_refs(
-                resource,
-                ReferenceCheck::REQUIRES_PROPERTY,
-                res_id,
-                &mut errors,
-            );
-            self.check_array_refs(
-                resource,
-                ReferenceCheck::RECOMMENDS_PROPERTY,
-                res_id,
-                &mut errors,
-            );
-            self.check_array_refs(
-                resource,
-                ReferenceCheck::SUBCLASS_OF_CLASS,
-                res_id,
-                &mut errors,
-            );
-        }
-
-        if is_property {
-            // class_types accepts BOTH Class and InductiveType IRIs
-            // (D32 §3.5): `data_type: core:resource(_array)` properties
-            // reference Classes; `data_type: core:inductive` properties
-            // reference InductiveTypes. Walk the array and accept
-            // either kind.
-            if let Some(value) = resource.get(&iri(wk::CLASS_TYPES)) {
-                for target in value.as_iri_array() {
-                    match self.layer.resolve(&target) {
-                        Some(t)
-                            if t.is_instance_of(&iri(wk::CLASS))
-                                || t.is_instance_of(&iri(wk::INDUCTIVE_TYPE)) => {}
-                        Some(_) => errors.push(ValidationError {
-                            resource_id: res_id.clone(),
-                            property: Some(iri(wk::CLASS_TYPES)),
-                            rule: ValidationRule::UnresolvedClassReference,
-                            message: format!(
-                                "class_types: '{target}' resolves to a resource that is not an instance of core:Class or core:InductiveType"
-                            ),
-                        }),
-                        None => errors.push(ValidationError {
-                            resource_id: res_id.clone(),
-                            property: Some(iri(wk::CLASS_TYPES)),
-                            rule: ValidationRule::UnresolvedClassReference,
-                            message: format!(
-                                "class_types: '{target}' does not resolve to any resource in the layer chain"
-                            ),
-                        }),
-                    }
-                }
-            }
-            // `data_type` is a single resource ref (not an array).
-            if let Some(value) = resource.get(&iri(wk::DATA_TYPE_PROP)) {
-                if let Some(target) = value_as_iri(value) {
-                    self.check_resolves_to(&target, ReferenceCheck::DATA_TYPE, res_id, &mut errors);
-                }
-            }
-        }
-        errors
     }
 
     /// Rule 15: Comorphism well-formedness (D14 §4.5 / §5).
@@ -1621,57 +431,6 @@ impl Validator {
                 message: format!(
                     "{}: '{target}' does not resolve to any resource in the layer chain",
                     typed_ref.field_label,
-                ),
-            }),
-        }
-    }
-
-    /// Walk an array-valued reference field on `resource` and verify
-    /// every element resolves to a resource of the expected class.
-    fn check_array_refs(
-        &self,
-        resource: &Resource,
-        check: ReferenceCheck<'static>,
-        res_id: &Option<Iri>,
-        errors: &mut Vec<ValidationError>,
-    ) {
-        let Some(value) = resource.get(&iri(check.field_iri)) else {
-            return;
-        };
-        for target in value.as_iri_array() {
-            self.check_resolves_to(&target, check, res_id, errors);
-        }
-    }
-
-    /// Verify a single referenced IRI resolves to a resource of the
-    /// expected class. Reports unresolved or wrong-kind references
-    /// against `ValidationRule::UnresolvedClassReference`.
-    fn check_resolves_to(
-        &self,
-        target: &Iri,
-        check: ReferenceCheck<'_>,
-        res_id: &Option<Iri>,
-        errors: &mut Vec<ValidationError>,
-    ) {
-        let expected_class = iri(check.expected_class_iri);
-        match self.layer.resolve(target) {
-            Some(target_resource) if target_resource.is_instance_of(&expected_class) => {}
-            Some(_) => errors.push(ValidationError {
-                resource_id: res_id.clone(),
-                property: Some(iri(check.field_iri)),
-                rule: ValidationRule::UnresolvedClassReference,
-                message: format!(
-                    "{}: '{target}' resolves to a resource that is not an instance of {}",
-                    check.field_label, check.expected_class_label,
-                ),
-            }),
-            None => errors.push(ValidationError {
-                resource_id: res_id.clone(),
-                property: Some(iri(check.field_iri)),
-                rule: ValidationRule::UnresolvedClassReference,
-                message: format!(
-                    "{}: '{target}' does not resolve to any resource in the layer chain",
-                    check.field_label,
                 ),
             }),
         }
@@ -2048,45 +807,10 @@ impl Validator {
     }
 }
 
-// --- Format validation helpers ---
-
-fn is_valid_date(s: &str) -> bool {
-    let re = regex::Regex::new(r"^\d{4}-\d{2}-\d{2}$").unwrap();
-    if !re.is_match(s) {
-        return false;
-    }
-    let parts: Vec<&str> = s.split('-').collect();
-    if parts.len() != 3 {
-        return false;
-    }
-    let year: u32 = parts[0].parse().unwrap_or(0);
-    let month: u32 = parts[1].parse().unwrap_or(0);
-    let day: u32 = parts[2].parse().unwrap_or(0);
-    (1..=9999).contains(&year) && (1..=12).contains(&month) && (1..=31).contains(&day)
-}
-
-fn is_valid_datetime(s: &str) -> bool {
-    // Accept ISO 8601 with timezone: YYYY-MM-DDTHH:MM:SSZ or +HH:MM
-    let re = regex::Regex::new(r"^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}(\.\d+)?(Z|[+-]\d{2}:\d{2})$")
-        .unwrap();
-    re.is_match(s)
-}
-
-fn is_valid_time(s: &str) -> bool {
-    let re = regex::Regex::new(r"^\d{2}:\d{2}:\d{2}(\.\d+)?(Z|[+-]\d{2}:\d{2})?$").unwrap();
-    re.is_match(s)
-}
-
-fn is_valid_uuid(s: &str) -> bool {
-    let re = regex::Regex::new(
-        r"^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
-    )
-    .unwrap();
-    re.is_match(s)
-}
+// --- Module-level shared helpers ---
 
 /// Helper: parse a well-known constant into an Iri.
-fn iri(s: &str) -> Iri {
+pub(crate) fn iri(s: &str) -> Iri {
     Iri::parse(s).expect("well-known IRI constants must be valid")
 }
 
@@ -2094,7 +818,7 @@ fn iri(s: &str) -> Iri {
 /// `Value::ResourceRef` (canonical) and `Value::String` (the JSON
 /// parser stores all strings as `Value::String` — `data_type` is
 /// frequently authored as a bare string in source ontologies).
-fn value_as_iri(value: &Value) -> Option<Iri> {
+pub(crate) fn value_as_iri(value: &Value) -> Option<Iri> {
     match value {
         Value::ResourceRef(i) => Some(i.clone()),
         Value::String(s) => Iri::parse(s).ok(),
@@ -2135,7 +859,7 @@ fn is_option_of_class(value: &Value, class_iri: &str) -> bool {
 
 /// Format a resource's `is_a` list for inclusion in an error message.
 /// `[]` for empty, otherwise `[a, b, c]`.
-fn format_is_a_list(classes: Vec<Iri>) -> String {
+pub(crate) fn format_is_a_list(classes: Vec<Iri>) -> String {
     if classes.is_empty() {
         "[]".to_string()
     } else {
@@ -2149,7 +873,7 @@ fn format_is_a_list(classes: Vec<Iri>) -> String {
 }
 
 /// Format an `&[&Iri]` slice for inclusion in an error message.
-fn format_iri_refs(refs: &[&Iri]) -> String {
+pub(crate) fn format_iri_refs(refs: &[&Iri]) -> String {
     let joined = refs
         .iter()
         .map(|i| i.as_str())
@@ -2173,66 +897,23 @@ struct ComorphismFormatRef<'a> {
     expected_label: &'a str,
 }
 
-/// Bundle of "what we're checking" for the class-definition reference
-/// validation pass (rule 14, eigenius#26). One value per field/expected-
-/// class pair; the constants below cover the five sites the validator
-/// inspects.
-#[derive(Copy, Clone)]
-struct ReferenceCheck<'a> {
-    /// Field IRI on the source resource (e.g. `core:requires`).
-    field_iri: &'a str,
-    /// Human label for the field used in error messages (e.g. `requires`).
-    field_label: &'a str,
-    /// Class IRI the referenced resource must be an instance of
-    /// (e.g. `core:Property`).
-    expected_class_iri: &'a str,
-    /// Human label for the expected class (e.g. `core:Property`).
-    expected_class_label: &'a str,
-}
-
-impl ReferenceCheck<'static> {
-    const REQUIRES_PROPERTY: Self = Self {
-        field_iri: wk::REQUIRES,
-        field_label: "requires",
-        expected_class_iri: wk::PROPERTY,
-        expected_class_label: "core:Property",
-    };
-    const RECOMMENDS_PROPERTY: Self = Self {
-        field_iri: wk::RECOMMENDS,
-        field_label: "recommends",
-        expected_class_iri: wk::PROPERTY,
-        expected_class_label: "core:Property",
-    };
-    const SUBCLASS_OF_CLASS: Self = Self {
-        field_iri: wk::PARENT_CLASSES,
-        field_label: "subclass_of",
-        expected_class_iri: wk::CLASS,
-        expected_class_label: "core:Class",
-    };
-    // class_types accepts both Class and InductiveType references
-    // (D32 §3.5); the check is open-coded in
-    // `check_class_definition_references` because `ReferenceCheck` is
-    // single-class-only.
-    const DATA_TYPE: Self = Self {
-        field_iri: wk::DATA_TYPE_PROP,
-        field_label: "data_type",
-        expected_class_iri: wk::DATA_TYPE,
-        expected_class_label: "core:DataType",
-    };
-}
-
 #[cfg(test)]
 mod tests {
+    //! Driver-loop and cross-rule integration tests, plus the test
+    //! helpers (`build_core_layer`, `make_resource`, …) re-used by
+    //! the per-rule test modules. Rule-specific test cases live
+    //! alongside their rule under `validation::rules::*`.
+
     use super::*;
     use crate::layer::LayerBuilder;
     use crate::ontology::eigon_json;
     use std::sync::Arc;
 
-    fn iri(s: &str) -> Iri {
+    pub(in crate::validation) fn iri(s: &str) -> Iri {
         Iri::parse(s).unwrap()
     }
 
-    fn make_resource(id: &str, props: Vec<(&str, Value)>) -> Resource {
+    pub(in crate::validation) fn make_resource(id: &str, props: Vec<(&str, Value)>) -> Resource {
         let mut r = Resource::new(iri(id));
         for (k, v) in props {
             r.set(iri(k), v);
@@ -2240,7 +921,7 @@ mod tests {
         r
     }
 
-    fn build_core_layer() -> Arc<Layer> {
+    pub(in crate::validation) fn build_core_layer() -> Arc<Layer> {
         let core_json = include_str!("../../../ontologies/core/core-ontology.json");
         let resources = eigon_json::parse_document(core_json).unwrap();
         let mut builder = LayerBuilder::new("core", None);
@@ -2295,214 +976,6 @@ mod tests {
     }
 
     #[test]
-    fn type_mismatch() {
-        let core = build_core_layer();
-        let mut builder = LayerBuilder::new("test", Some(core));
-        // description should be a string, give it an integer
-        builder
-            .add_resource(make_resource(
-                "urn:eigenius:test:bad_type",
-                vec![
-                    (
-                        wk::IS_A,
-                        Value::Array(vec![Value::String(wk::CLASS.to_string())]),
-                    ),
-                    (wk::DESCRIPTION, Value::Integer(42)), // Wrong type!
-                    (wk::SHORT_NAME, Value::String("bad".into())),
-                ],
-            ))
-            .unwrap();
-        let layer = Arc::new(builder.build(crate::layer::LayerStorage::in_memory()));
-
-        let validator = Validator::new(Arc::clone(&layer));
-        let errors = validator.validate();
-        assert!(
-            errors
-                .iter()
-                .any(|e| e.rule == ValidationRule::TypeMismatch),
-            "expected TypeMismatch error; got: {errors:?}"
-        );
-    }
-
-    #[test]
-    fn allows_only_violation() {
-        let core = build_core_layer();
-        let mut builder = LayerBuilder::new("test", Some(core));
-        // data_type has allows_only constraint — use an invalid value
-        builder
-            .add_resource(make_resource(
-                "urn:eigenius:test:bad_dt",
-                vec![
-                    (
-                        wk::IS_A,
-                        Value::Array(vec![Value::String(wk::PROPERTY.to_string())]),
-                    ),
-                    (wk::DESCRIPTION, Value::String("bad data_type".into())),
-                    (wk::SHORT_NAME, Value::String("bad_dt".into())),
-                    (
-                        wk::DATA_TYPE_PROP,
-                        Value::String("urn:eigenius:core:nonexistent".to_string()),
-                    ),
-                ],
-            ))
-            .unwrap();
-        let layer = Arc::new(builder.build(crate::layer::LayerStorage::in_memory()));
-
-        let validator = Validator::new(Arc::clone(&layer));
-        let errors = validator.validate();
-        assert!(errors
-            .iter()
-            .any(|e| e.rule == ValidationRule::AllowedValueViolation));
-    }
-
-    #[test]
-    fn domain_violation() {
-        let core = build_core_layer();
-        let mut builder = LayerBuilder::new("test", Some(core));
-        // 'requires' has domain [Class], but we put it on a non-Class resource
-        builder
-            .add_resource(make_resource(
-                "urn:eigenius:test:not_a_class",
-                vec![
-                    (
-                        wk::IS_A,
-                        Value::Array(vec![Value::String(wk::PROPERTY.to_string())]),
-                    ),
-                    (wk::DESCRIPTION, Value::String("I'm a property".into())),
-                    (wk::SHORT_NAME, Value::String("not_a_class".into())),
-                    (wk::DATA_TYPE_PROP, Value::String(wk::STRING.to_string())),
-                    // 'requires' is Class-only, but this is a Property
-                    (
-                        wk::REQUIRES,
-                        Value::Array(vec![Value::String("urn:eigenius:test:foo".to_string())]),
-                    ),
-                ],
-            ))
-            .unwrap();
-        let layer = Arc::new(builder.build(crate::layer::LayerStorage::in_memory()));
-
-        let validator = Validator::new(Arc::clone(&layer));
-        let errors = validator.validate();
-        assert!(errors
-            .iter()
-            .any(|e| e.rule == ValidationRule::DomainViolation));
-    }
-
-    #[test]
-    fn inheritance_requires() {
-        let core = build_core_layer();
-        let mut builder = LayerBuilder::new("test", Some(core));
-
-        // Define Animal class requiring 'name'
-        builder
-            .add_resource(make_resource(
-                "urn:eigenius:test:Animal",
-                vec![
-                    (
-                        wk::IS_A,
-                        Value::Array(vec![Value::String(wk::CLASS.to_string())]),
-                    ),
-                    (wk::DESCRIPTION, Value::String("An animal".into())),
-                    (wk::SHORT_NAME, Value::String("Animal".into())),
-                    (
-                        wk::REQUIRES,
-                        Value::Array(vec![Value::String("urn:eigenius:test:name".to_string())]),
-                    ),
-                ],
-            ))
-            .unwrap();
-
-        // Define Dog class extending Animal, requiring 'breed'
-        builder
-            .add_resource(make_resource(
-                "urn:eigenius:test:Dog",
-                vec![
-                    (
-                        wk::IS_A,
-                        Value::Array(vec![Value::String(wk::CLASS.to_string())]),
-                    ),
-                    (wk::DESCRIPTION, Value::String("A dog".into())),
-                    (wk::SHORT_NAME, Value::String("Dog".into())),
-                    (
-                        wk::PARENT_CLASSES,
-                        Value::Array(vec![Value::String("urn:eigenius:test:Animal".to_string())]),
-                    ),
-                    (
-                        wk::REQUIRES,
-                        Value::Array(vec![Value::String("urn:eigenius:test:breed".to_string())]),
-                    ),
-                ],
-            ))
-            .unwrap();
-
-        // Define the properties
-        builder
-            .add_resource(make_resource(
-                "urn:eigenius:test:name",
-                vec![
-                    (
-                        wk::IS_A,
-                        Value::Array(vec![Value::String(wk::PROPERTY.to_string())]),
-                    ),
-                    (wk::DESCRIPTION, Value::String("Name".into())),
-                    (wk::SHORT_NAME, Value::String("name".into())),
-                    (wk::DATA_TYPE_PROP, Value::String(wk::STRING.to_string())),
-                ],
-            ))
-            .unwrap();
-
-        builder
-            .add_resource(make_resource(
-                "urn:eigenius:test:breed",
-                vec![
-                    (
-                        wk::IS_A,
-                        Value::Array(vec![Value::String(wk::PROPERTY.to_string())]),
-                    ),
-                    (wk::DESCRIPTION, Value::String("Breed".into())),
-                    (wk::SHORT_NAME, Value::String("breed".into())),
-                    (wk::DATA_TYPE_PROP, Value::String(wk::STRING.to_string())),
-                ],
-            ))
-            .unwrap();
-
-        // Create a Dog instance missing 'name' (inherited from Animal)
-        builder
-            .add_resource(make_resource(
-                "urn:eigenius:test:rex",
-                vec![
-                    (
-                        wk::IS_A,
-                        Value::Array(vec![Value::String("urn:eigenius:test:Dog".to_string())]),
-                    ),
-                    (
-                        "urn:eigenius:test:breed",
-                        Value::String("German Shepherd".into()),
-                    ),
-                    // Missing 'name'!
-                ],
-            ))
-            .unwrap();
-
-        let layer = Arc::new(builder.build(crate::layer::LayerStorage::in_memory()));
-        let validator = Validator::new(Arc::clone(&layer));
-        let errors = validator.validate();
-
-        // Should have a MissingRequired for 'name' on rex
-        let rex_errors: Vec<_> = errors
-            .iter()
-            .filter(|e| {
-                e.resource_id.as_ref().map(|i| i.as_str()) == Some("urn:eigenius:test:rex")
-                    && e.rule == ValidationRule::MissingRequired
-            })
-            .collect();
-        assert!(
-            !rex_errors.is_empty(),
-            "Dog instance missing inherited 'name' should fail validation"
-        );
-    }
-
-    #[test]
     fn valid_resource_passes() {
         let core = build_core_layer();
         let mut builder = LayerBuilder::new("test", Some(core));
@@ -2532,26 +1005,6 @@ mod tests {
             })
             .collect();
         assert!(my_errors.is_empty(), "valid resource should have no errors");
-    }
-
-    #[test]
-    fn format_date_validation() {
-        assert!(is_valid_date("2026-04-11"));
-        assert!(!is_valid_date("2026-13-01"));
-        assert!(!is_valid_date("not-a-date"));
-    }
-
-    #[test]
-    fn format_datetime_validation() {
-        assert!(is_valid_datetime("2026-04-11T14:30:00Z"));
-        assert!(is_valid_datetime("2026-04-11T14:30:00+05:30"));
-        assert!(!is_valid_datetime("2026-04-11"));
-    }
-
-    #[test]
-    fn format_uuid_validation() {
-        assert!(is_valid_uuid("550e8400-e29b-41d4-a716-446655440000"));
-        assert!(!is_valid_uuid("not-a-uuid"));
     }
 
     // --- Epistemic base class validation (Step 6) ---
@@ -3087,148 +1540,8 @@ mod tests {
         );
     }
 
-    // --- eigenius#26: class-definition reference integrity ---
+    // --- Comorphism well-formedness tests (Rule 15, D14 §4.5 / §5) ---
 
-    /// A class that `requires` an IRI with no matching Property
-    /// declaration anywhere in the chain must fail validation rather
-    /// than commit cleanly.
-    #[test]
-    fn class_requires_unresolved_property_is_rejected() {
-        let core = build_core_layer();
-        let mut top = LayerBuilder::new("test", Some(core));
-
-        // Class declaring a requires reference to a property that
-        // doesn't exist (typo / forgotten-declaration scenario).
-        let bad_class = make_resource(
-            "urn:eigenius:test:Foo",
-            vec![
-                (
-                    wk::IS_A,
-                    Value::Array(vec![Value::String(wk::CLASS.into())]),
-                ),
-                (wk::SHORT_NAME, Value::String("Foo".into())),
-                (wk::DESCRIPTION, Value::String("Test class.".into())),
-                (
-                    wk::REQUIRES,
-                    Value::Array(vec![Value::String(
-                        "urn:eigenius:test:totally_made_up_property".into(),
-                    )]),
-                ),
-            ],
-        );
-        top.add_resource(bad_class).unwrap();
-        let layer = Arc::new(top.build(crate::layer::LayerStorage::in_memory()));
-
-        let validator = Validator::new(Arc::clone(&layer));
-        let errors = validator.validate();
-        let dangling: Vec<_> = errors
-            .iter()
-            .filter(|e| {
-                e.rule == ValidationRule::UnresolvedClassReference
-                    && e.message.contains("totally_made_up_property")
-            })
-            .collect();
-        assert_eq!(
-            dangling.len(),
-            1,
-            "expected exactly one UnresolvedClassReference for the missing property; got {errors:?}"
-        );
-    }
-
-    /// Same class, but the referenced property is declared in the
-    /// same load batch — must validate cleanly.
-    #[test]
-    fn class_requires_same_batch_property_is_accepted() {
-        let core = build_core_layer();
-        let mut top = LayerBuilder::new("test", Some(core));
-
-        let prop = make_resource(
-            "urn:eigenius:test:my_prop",
-            vec![
-                (
-                    wk::IS_A,
-                    Value::Array(vec![Value::String(wk::PROPERTY.into())]),
-                ),
-                (wk::SHORT_NAME, Value::String("my_prop".into())),
-                (wk::DESCRIPTION, Value::String("A test property.".into())),
-                (
-                    wk::DATA_TYPE_PROP,
-                    Value::String("urn:eigenius:core:string".into()),
-                ),
-            ],
-        );
-        let class = make_resource(
-            "urn:eigenius:test:Foo",
-            vec![
-                (
-                    wk::IS_A,
-                    Value::Array(vec![Value::String(wk::CLASS.into())]),
-                ),
-                (wk::SHORT_NAME, Value::String("Foo".into())),
-                (wk::DESCRIPTION, Value::String("Test class.".into())),
-                (
-                    wk::REQUIRES,
-                    Value::Array(vec![Value::String("urn:eigenius:test:my_prop".into())]),
-                ),
-            ],
-        );
-        top.add_resource(prop).unwrap();
-        top.add_resource(class).unwrap();
-        let layer = Arc::new(top.build(crate::layer::LayerStorage::in_memory()));
-
-        let validator = Validator::new(Arc::clone(&layer));
-        let errors = validator.validate();
-        let dangling: Vec<_> = errors
-            .iter()
-            .filter(|e| e.rule == ValidationRule::UnresolvedClassReference)
-            .collect();
-        assert!(
-            dangling.is_empty(),
-            "valid forward-reference should not surface UnresolvedClassReference; got {dangling:?}"
-        );
-    }
-
-    /// A property whose `data_type` doesn't resolve must fail.
-    #[test]
-    fn property_data_type_unresolved_is_rejected() {
-        let core = build_core_layer();
-        let mut top = LayerBuilder::new("test", Some(core));
-
-        let bad_prop = make_resource(
-            "urn:eigenius:test:my_prop",
-            vec![
-                (
-                    wk::IS_A,
-                    Value::Array(vec![Value::String(wk::PROPERTY.into())]),
-                ),
-                (wk::SHORT_NAME, Value::String("my_prop".into())),
-                (wk::DESCRIPTION, Value::String("Bad prop.".into())),
-                (
-                    wk::DATA_TYPE_PROP,
-                    Value::String("urn:eigenius:test:not_a_real_type".into()),
-                ),
-            ],
-        );
-        top.add_resource(bad_prop).unwrap();
-        let layer = Arc::new(top.build(crate::layer::LayerStorage::in_memory()));
-
-        let validator = Validator::new(Arc::clone(&layer));
-        let errors = validator.validate();
-        let dangling: Vec<_> = errors
-            .iter()
-            .filter(|e| {
-                e.rule == ValidationRule::UnresolvedClassReference
-                    && e.message.contains("not_a_real_type")
-            })
-            .collect();
-        assert_eq!(
-            dangling.len(),
-            1,
-            "expected exactly one UnresolvedClassReference for the missing data_type; got {errors:?}"
-        );
-    }
-
-    /// A property's `class_types` referencing a non-Class fails.
     /// Build a layer chain containing the bootstrap ontologies (core +
     /// institution + program + reflection + notebook). Used by tests
     /// that exercise the Comorphism well-formedness rule (D14 §4.5)
@@ -3506,846 +1819,6 @@ mod tests {
             dangling.len(),
             1,
             "expected one UnresolvedClassReference for unresolvable transformation; got {errors:?}"
-        );
-    }
-
-    #[test]
-    fn property_class_types_pointing_at_non_class_is_rejected() {
-        let core = build_core_layer();
-        let mut top = LayerBuilder::new("test", Some(core));
-
-        // A non-Class resource (just an instance of `core:Class`'s
-        // base — actually use core:DataType, which is a Class but its
-        // *instances* aren't classes themselves).
-        let instance = make_resource(
-            "urn:eigenius:test:not_a_class",
-            vec![
-                // is_a a DataType, NOT a Class.
-                (
-                    wk::IS_A,
-                    Value::Array(vec![Value::String(wk::DATA_TYPE.into())]),
-                ),
-                (wk::SHORT_NAME, Value::String("not_a_class".into())),
-                (wk::DESCRIPTION, Value::String("placeholder".into())),
-            ],
-        );
-        let bad_prop = make_resource(
-            "urn:eigenius:test:my_prop",
-            vec![
-                (
-                    wk::IS_A,
-                    Value::Array(vec![Value::String(wk::PROPERTY.into())]),
-                ),
-                (wk::SHORT_NAME, Value::String("my_prop".into())),
-                (wk::DESCRIPTION, Value::String("Bad prop.".into())),
-                (
-                    wk::DATA_TYPE_PROP,
-                    Value::String("urn:eigenius:core:resource".into()),
-                ),
-                (
-                    wk::CLASS_TYPES,
-                    Value::Array(vec![Value::String("urn:eigenius:test:not_a_class".into())]),
-                ),
-            ],
-        );
-        top.add_resource(instance).unwrap();
-        top.add_resource(bad_prop).unwrap();
-        let layer = Arc::new(top.build(crate::layer::LayerStorage::in_memory()));
-
-        let validator = Validator::new(Arc::clone(&layer));
-        let errors = validator.validate();
-        let dangling: Vec<_> = errors
-            .iter()
-            .filter(|e| {
-                e.rule == ValidationRule::UnresolvedClassReference
-                    && e.message.contains("not_a_class")
-            })
-            .collect();
-        assert_eq!(
-            dangling.len(),
-            1,
-            "expected one UnresolvedClassReference for class_types pointing at a non-Class; got {errors:?}"
-        );
-    }
-
-    // ──────────────────────────────────────────────────────────────────
-    // Inductive value validation — D32 §3.5
-    // ──────────────────────────────────────────────────────────────────
-
-    /// Build a minimal `Nat = zero | succ(Nat)` ontology layer + a
-    /// property `nat_value : core:inductive` with `class_types: [Nat]`,
-    /// returning the chain layer ready to commit Nat values against.
-    fn build_nat_layer() -> Arc<Layer> {
-        let core = build_core_layer();
-        let mut builder = LayerBuilder::new("test_nat", Some(core));
-
-        // ctor `zero`: no args.
-        let zero_ctor = make_resource(
-            "urn:eigenius:test:Nat:zero",
-            vec![
-                (
-                    wk::IS_A,
-                    Value::Array(vec![Value::ResourceRef(iri(wk::INDUCTIVE_CTOR))]),
-                ),
-                (wk::CTOR_NAME, Value::String("zero".into())),
-                (wk::ARG_TYPES, Value::Array(vec![])),
-            ],
-        );
-
-        // ctor `succ(pred: Nat)`.
-        let succ_arg = make_resource(
-            "urn:eigenius:test:Nat:succ:pred",
-            vec![
-                (
-                    wk::IS_A,
-                    Value::Array(vec![Value::ResourceRef(iri(wk::INDUCTIVE_ARG_TYPE))]),
-                ),
-                (wk::ARG_NAME, Value::String("pred".into())),
-                (wk::TYPE_NAME, Value::String("urn:eigenius:test:Nat".into())),
-            ],
-        );
-        let succ_ctor = make_resource(
-            "urn:eigenius:test:Nat:succ",
-            vec![
-                (
-                    wk::IS_A,
-                    Value::Array(vec![Value::ResourceRef(iri(wk::INDUCTIVE_CTOR))]),
-                ),
-                (wk::CTOR_NAME, Value::String("succ".into())),
-                (
-                    wk::ARG_TYPES,
-                    Value::Array(vec![Value::Embedded(Box::new(succ_arg))]),
-                ),
-            ],
-        );
-
-        let nat = make_resource(
-            "urn:eigenius:test:Nat",
-            vec![
-                (
-                    wk::IS_A,
-                    Value::Array(vec![Value::ResourceRef(iri(wk::INDUCTIVE_TYPE))]),
-                ),
-                (wk::SHORT_NAME, Value::String("Nat".into())),
-                (
-                    wk::CTORS,
-                    Value::Array(vec![
-                        Value::Embedded(Box::new(zero_ctor)),
-                        Value::Embedded(Box::new(succ_ctor)),
-                    ]),
-                ),
-            ],
-        );
-
-        // Property `nat_value : core:inductive` typed at Nat.
-        let nat_value_prop = make_resource(
-            "urn:eigenius:test:nat_value",
-            vec![
-                (
-                    wk::IS_A,
-                    Value::Array(vec![Value::ResourceRef(iri(wk::PROPERTY))]),
-                ),
-                (wk::SHORT_NAME, Value::String("nat_value".into())),
-                (wk::DATA_TYPE_PROP, Value::ResourceRef(iri(wk::INDUCTIVE))),
-                (
-                    wk::CLASS_TYPES,
-                    Value::Array(vec![Value::ResourceRef(iri("urn:eigenius:test:Nat"))]),
-                ),
-            ],
-        );
-
-        builder.add_resource(nat).unwrap();
-        builder.add_resource(nat_value_prop).unwrap();
-        Arc::new(builder.build(crate::layer::LayerStorage::in_memory()))
-    }
-
-    /// `succ(succ(zero))` as a JSON tagged-dict tree.
-    fn nat_succ_succ_zero() -> serde_json::Value {
-        serde_json::json!({
-            "ctor": "succ",
-            "args": [{
-                "ctor": "succ",
-                "args": [{
-                    "ctor": "zero",
-                    "args": []
-                }]
-            }]
-        })
-    }
-
-    #[test]
-    fn inductive_value_validates_succ_succ_zero() {
-        let nat_layer = build_nat_layer();
-        let mut top = LayerBuilder::new("test_top", Some(nat_layer));
-
-        let holder = make_resource(
-            "urn:eigenius:test:n2",
-            vec![(
-                "urn:eigenius:test:nat_value",
-                Value::Json(nat_succ_succ_zero()),
-            )],
-        );
-        top.add_resource(holder).unwrap();
-        let layer = Arc::new(top.build(crate::layer::LayerStorage::in_memory()));
-
-        let validator = Validator::new(Arc::clone(&layer));
-        let errors = validator.validate();
-        let inductive_errors: Vec<_> = errors
-            .iter()
-            .filter(|e| e.rule == ValidationRule::InductiveValueMismatch)
-            .collect();
-        assert!(
-            inductive_errors.is_empty(),
-            "expected no InductiveValueMismatch on succ(succ(zero)); got {errors:?}"
-        );
-    }
-
-    #[test]
-    fn inductive_value_rejects_unknown_ctor() {
-        let nat_layer = build_nat_layer();
-        let mut top = LayerBuilder::new("test_top", Some(nat_layer));
-
-        let bad = make_resource(
-            "urn:eigenius:test:bad",
-            vec![(
-                "urn:eigenius:test:nat_value",
-                Value::Json(serde_json::json!({
-                    "ctor": "infinity",
-                    "args": []
-                })),
-            )],
-        );
-        top.add_resource(bad).unwrap();
-        let layer = Arc::new(top.build(crate::layer::LayerStorage::in_memory()));
-
-        let errors = Validator::new(Arc::clone(&layer)).validate();
-        let mismatches: Vec<_> = errors
-            .iter()
-            .filter(|e| e.rule == ValidationRule::InductiveValueMismatch)
-            .collect();
-        assert_eq!(
-            mismatches.len(),
-            1,
-            "expected exactly one InductiveValueMismatch for unknown ctor; got {errors:?}"
-        );
-        assert!(
-            mismatches[0].message.contains("infinity"),
-            "error must mention the offending ctor name: {}",
-            mismatches[0].message
-        );
-    }
-
-    #[test]
-    fn inductive_value_rejects_arity_mismatch() {
-        let nat_layer = build_nat_layer();
-        let mut top = LayerBuilder::new("test_top", Some(nat_layer));
-
-        // succ takes one arg; supply zero.
-        let bad = make_resource(
-            "urn:eigenius:test:bad_arity",
-            vec![(
-                "urn:eigenius:test:nat_value",
-                Value::Json(serde_json::json!({
-                    "ctor": "succ",
-                    "args": []
-                })),
-            )],
-        );
-        top.add_resource(bad).unwrap();
-        let layer = Arc::new(top.build(crate::layer::LayerStorage::in_memory()));
-
-        let errors = Validator::new(Arc::clone(&layer)).validate();
-        let mismatches: Vec<_> = errors
-            .iter()
-            .filter(|e| e.rule == ValidationRule::InductiveValueMismatch)
-            .collect();
-        assert_eq!(
-            mismatches.len(),
-            1,
-            "expected one InductiveValueMismatch for arity mismatch; got {errors:?}"
-        );
-        assert!(
-            mismatches[0].message.contains("expects 1 arg"),
-            "error must describe the arity mismatch: {}",
-            mismatches[0].message
-        );
-    }
-
-    // ──────────────────────────────────────────────────────────────────
-    // FormulaTerm App-spine arity check — D32 §5.4 / Phase 19d.0.d
-    // ──────────────────────────────────────────────────────────────────
-
-    /// Build a layer chain rooted at the embedded core+formulas
-    /// ontologies plus a property `formula_value : core:inductive`
-    /// typed at FormulaTerm. Used by the arity-check tests.
-    fn build_formula_layer() -> Arc<Layer> {
-        // Reuse bootstrap so the formulas: layer (with FormulaTerm +
-        // operator catalog) sits in the chain.
-        let ctx = crate::bootstrap::bootstrap().expect("bootstrap with formulas layer");
-        let formulas = Arc::clone(ctx.head().parent().expect("notebook has parent"));
-
-        let mut builder = LayerBuilder::new("test_formula", Some(formulas));
-        let prop = make_resource(
-            "urn:eigenius:test:formula_value",
-            vec![
-                (
-                    wk::IS_A,
-                    Value::Array(vec![Value::ResourceRef(iri(wk::PROPERTY))]),
-                ),
-                (wk::SHORT_NAME, Value::String("formula_value".into())),
-                (wk::DATA_TYPE_PROP, Value::ResourceRef(iri(wk::INDUCTIVE))),
-                (
-                    wk::CLASS_TYPES,
-                    Value::Array(vec![Value::ResourceRef(iri(
-                        "urn:eigenius:formulas:FormulaTerm",
-                    ))]),
-                ),
-            ],
-        );
-        builder.add_resource(prop).unwrap();
-        Arc::new(builder.build(crate::layer::LayerStorage::in_memory()))
-    }
-
-    /// `App(App(OpRef("formulas:ops:add"), Var("x")), LitFloat(2.0))`
-    /// — well-formed binary `add` invocation.
-    fn add_x_2() -> serde_json::Value {
-        serde_json::json!({
-            "ctor": "App",
-            "args": [
-                {
-                    "ctor": "App",
-                    "args": [
-                        {"ctor": "OpRef", "args": ["urn:eigenius:formulas:ops:add"]},
-                        {"ctor": "Var", "args": ["x"]}
-                    ]
-                },
-                {"ctor": "LitFloat", "args": [2.0]}
-            ]
-        })
-    }
-
-    #[test]
-    fn formula_term_well_formed_app_validates() {
-        let layer = build_formula_layer();
-        let mut top = LayerBuilder::new("test_top", Some(layer));
-        let holder = make_resource(
-            "urn:eigenius:test:f1",
-            vec![("urn:eigenius:test:formula_value", Value::Json(add_x_2()))],
-        );
-        top.add_resource(holder).unwrap();
-        let layer = Arc::new(top.build(crate::layer::LayerStorage::in_memory()));
-
-        let errors = Validator::new(Arc::clone(&layer)).validate();
-        let arity_errors: Vec<_> = errors
-            .iter()
-            .filter(|e| e.rule == ValidationRule::OperatorArityMismatch)
-            .collect();
-        assert!(
-            arity_errors.is_empty(),
-            "well-formed `add(x, 2)` must not raise OperatorArityMismatch; got {arity_errors:?}"
-        );
-    }
-
-    #[test]
-    fn formula_term_app_rejects_arity_short() {
-        // `App(OpRef("add"), x)` — missing the second add argument.
-        let underapplied = serde_json::json!({
-            "ctor": "App",
-            "args": [
-                {"ctor": "OpRef", "args": ["urn:eigenius:formulas:ops:add"]},
-                {"ctor": "Var", "args": ["x"]}
-            ]
-        });
-
-        let layer = build_formula_layer();
-        let mut top = LayerBuilder::new("test_top", Some(layer));
-        let holder = make_resource(
-            "urn:eigenius:test:bad_arity",
-            vec![("urn:eigenius:test:formula_value", Value::Json(underapplied))],
-        );
-        top.add_resource(holder).unwrap();
-        let layer = Arc::new(top.build(crate::layer::LayerStorage::in_memory()));
-
-        let errors = Validator::new(Arc::clone(&layer)).validate();
-        let arity_errors: Vec<_> = errors
-            .iter()
-            .filter(|e| e.rule == ValidationRule::OperatorArityMismatch)
-            .collect();
-        assert_eq!(
-            arity_errors.len(),
-            1,
-            "expected one OperatorArityMismatch for under-applied `add`; got {errors:?}"
-        );
-        assert!(
-            arity_errors[0].message.contains("formulas:ops:add"),
-            "error must name the offending operator: {}",
-            arity_errors[0].message
-        );
-        assert!(
-            arity_errors[0].message.contains("arity 2"),
-            "error must mention the declared arity: {}",
-            arity_errors[0].message
-        );
-    }
-
-    #[test]
-    fn formula_term_app_rejects_arity_long() {
-        // `App(App(App(OpRef("neg"), x), y), z)` — `neg` is unary;
-        // the spine supplies three args.
-        let overapplied = serde_json::json!({
-            "ctor": "App",
-            "args": [
-                {
-                    "ctor": "App",
-                    "args": [
-                        {
-                            "ctor": "App",
-                            "args": [
-                                {"ctor": "OpRef", "args": ["urn:eigenius:formulas:ops:neg"]},
-                                {"ctor": "Var", "args": ["x"]}
-                            ]
-                        },
-                        {"ctor": "Var", "args": ["y"]}
-                    ]
-                },
-                {"ctor": "Var", "args": ["z"]}
-            ]
-        });
-
-        let layer = build_formula_layer();
-        let mut top = LayerBuilder::new("test_top", Some(layer));
-        let holder = make_resource(
-            "urn:eigenius:test:bad_arity_long",
-            vec![("urn:eigenius:test:formula_value", Value::Json(overapplied))],
-        );
-        top.add_resource(holder).unwrap();
-        let layer = Arc::new(top.build(crate::layer::LayerStorage::in_memory()));
-
-        let errors = Validator::new(Arc::clone(&layer)).validate();
-        let arity_errors: Vec<_> = errors
-            .iter()
-            .filter(|e| e.rule == ValidationRule::OperatorArityMismatch)
-            .collect();
-        assert!(
-            !arity_errors.is_empty(),
-            "expected an OperatorArityMismatch for over-applied `neg`; got {errors:?}"
-        );
-    }
-
-    // ──────────────────────────────────────────────────────────────────
-    // Phase 20a.2 — D40 chain-mirrored Lean expressions
-    // ──────────────────────────────────────────────────────────────────
-
-    /// Build a layer chain rooted at the embedded bootstrap (which now
-    /// carries `lean:LeanExpr` + siblings per D40) plus a property
-    /// `proposition_value : core:inductive` typed at `lean:LeanExpr`.
-    /// The chain looks like: core → program → reflection → institution
-    /// → runtime → formulas → lean-expressions → <this test layer>.
-    fn build_lean_expr_layer() -> Arc<Layer> {
-        let ctx = crate::bootstrap::bootstrap().expect("bootstrap with lean-expressions layer");
-        // After Phase 20a.4 the chain is:
-        //   notebook → lean-institution → lean-expressions → formulas → …
-        // We anchor at lean-institution (`ctx.head().parent()`) so the
-        // test layer has both the lean:LeanExpr InductiveTypes
-        // (resolved through `lean-expressions`) and the
-        // institution-side classes (LeanProofTerm etc.) reachable —
-        // notebook would also work, but anchoring above it keeps the
-        // chain focused.
-        let lean_layer = Arc::clone(
-            ctx.head()
-                .parent()
-                .expect("head has lean-institution parent"),
-        );
-
-        let mut builder = LayerBuilder::new("test_lean_expr", Some(lean_layer));
-        let prop = make_resource(
-            "urn:eigenius:test:proposition_value",
-            vec![
-                (
-                    wk::IS_A,
-                    Value::Array(vec![Value::ResourceRef(iri(wk::PROPERTY))]),
-                ),
-                (wk::SHORT_NAME, Value::String("proposition_value".into())),
-                (wk::DATA_TYPE_PROP, Value::ResourceRef(iri(wk::INDUCTIVE))),
-                (
-                    wk::CLASS_TYPES,
-                    Value::Array(vec![Value::ResourceRef(iri("urn:eigenius:lean:LeanExpr"))]),
-                ),
-            ],
-        );
-        builder.add_resource(prop).unwrap();
-        Arc::new(builder.build(crate::layer::LayerStorage::in_memory()))
-    }
-
-    /// `Lambda { binder_name = Str(Anon, "x"), binder_style = "default",
-    ///           binder_type = Const(Str(Anon, "Nat"), Nil),
-    ///           body = Var(0) }`
-    /// ≈ `λ x : Nat, x` — the smallest non-trivial closed Lean term.
-    fn lambda_x_in_nat() -> serde_json::Value {
-        let anon = serde_json::json!({"ctor": "Anon"});
-        let name_x = serde_json::json!({
-            "ctor": "Str",
-            "args": [anon.clone(), "x"]
-        });
-        let name_nat = serde_json::json!({
-            "ctor": "Str",
-            "args": [anon.clone(), "Nat"]
-        });
-        let nil = serde_json::json!({"ctor": "Nil"});
-        serde_json::json!({
-            "ctor": "Lambda",
-            "args": [
-                name_x,
-                "default",
-                {
-                    "ctor": "Const",
-                    "args": [name_nat, nil]
-                },
-                {"ctor": "Var", "args": [0]}
-            ]
-        })
-    }
-
-    #[test]
-    fn lean_expr_lambda_x_in_nat_validates() {
-        // Phase 20a.2 acceptance test: a hand-encoded `λ x : Nat, x`
-        // value commits cleanly against the chain-mirrored LeanExpr
-        // ontology — the chain-side type-check walks the tagged-dict
-        // shape, dispatches on each ctor name, and recurses into
-        // child arguments through LeanName / LeanLevelList / LeanExpr.
-        // No validator errors means the inductive-value walker
-        // successfully resolved every cross-reference and the
-        // ontology layer is structurally consistent.
-        let layer = build_lean_expr_layer();
-        let mut top = LayerBuilder::new("test_top", Some(layer));
-        let holder = make_resource(
-            "urn:eigenius:test:p1",
-            vec![(
-                "urn:eigenius:test:proposition_value",
-                Value::Json(lambda_x_in_nat()),
-            )],
-        );
-        top.add_resource(holder).unwrap();
-        let layer = Arc::new(top.build(crate::layer::LayerStorage::in_memory()));
-
-        let errors = Validator::new(Arc::clone(&layer)).validate();
-        let inductive_errors: Vec<_> = errors
-            .iter()
-            .filter(|e| e.rule == ValidationRule::InductiveValueMismatch)
-            .collect();
-        assert!(
-            inductive_errors.is_empty(),
-            "well-formed `λ x : Nat, x` must validate as a lean:LeanExpr; got {inductive_errors:?}"
-        );
-    }
-
-    #[test]
-    fn lean_expr_unknown_ctor_rejected() {
-        // A value with a ctor name the LeanExpr inductive doesn't
-        // declare must surface as `InductiveValueMismatch` — exercises
-        // the per-ctor lookup in `walk_inductive_value`.
-        let bogus = serde_json::json!({
-            "ctor": "MetaVar",
-            "args": [42]
-        });
-        let layer = build_lean_expr_layer();
-        let mut top = LayerBuilder::new("test_top", Some(layer));
-        let holder = make_resource(
-            "urn:eigenius:test:bad_ctor",
-            vec![("urn:eigenius:test:proposition_value", Value::Json(bogus))],
-        );
-        top.add_resource(holder).unwrap();
-        let layer = Arc::new(top.build(crate::layer::LayerStorage::in_memory()));
-
-        let errors = Validator::new(Arc::clone(&layer)).validate();
-        let inductive_errors: Vec<_> = errors
-            .iter()
-            .filter(|e| e.rule == ValidationRule::InductiveValueMismatch)
-            .collect();
-        assert!(
-            !inductive_errors.is_empty(),
-            "unknown ctor `MetaVar` must trigger an InductiveValueMismatch; got {errors:?}"
-        );
-    }
-
-    #[test]
-    fn lean_expr_resolves_lean_layer_inductives() {
-        // Sanity check: after bootstrap, every LeanExpr-related
-        // InductiveType is reachable from the head as
-        // `is_instance_of(core:InductiveType)`. Catches typos in the
-        // ontology JSON or missing entries in the layer chain.
-        let ctx = crate::bootstrap::bootstrap().expect("bootstrap");
-        let head = Arc::clone(ctx.head());
-        for ind_iri in &[
-            "urn:eigenius:lean:LeanName",
-            "urn:eigenius:lean:LeanLevel",
-            "urn:eigenius:lean:LeanLevelList",
-            "urn:eigenius:lean:LeanExpr",
-        ] {
-            let parsed = iri(ind_iri);
-            let resolved = head.resolve(&parsed).unwrap_or_else(|| {
-                panic!("`{ind_iri}` should resolve from the bootstrap chain head")
-            });
-            assert!(
-                resolved.is_instance_of(&iri(wk::INDUCTIVE_TYPE)),
-                "`{ind_iri}` should be an InductiveType"
-            );
-        }
-    }
-
-    /// Build a layer with a property `proposition_value` whose
-    /// `data_type` is the caller-supplied IRI and whose `class_types`
-    /// references `lean:LeanExpr`. Powers the Option A tests that
-    /// exercise `core:resource` / `core:resource_array` carrying
-    /// inductive values without going through `core:inductive`.
-    fn build_lean_expr_property_layer(data_type_iri: &str) -> Arc<Layer> {
-        let ctx = crate::bootstrap::bootstrap().expect("bootstrap with lean-expressions layer");
-        let lean_layer = Arc::clone(
-            ctx.head()
-                .parent()
-                .expect("head has lean-institution parent"),
-        );
-
-        let mut builder = LayerBuilder::new("test_lean_expr_resource", Some(lean_layer));
-        let prop = make_resource(
-            "urn:eigenius:test:proposition_value",
-            vec![
-                (
-                    wk::IS_A,
-                    Value::Array(vec![Value::ResourceRef(iri(wk::PROPERTY))]),
-                ),
-                (wk::SHORT_NAME, Value::String("proposition_value".into())),
-                (wk::DATA_TYPE_PROP, Value::ResourceRef(iri(data_type_iri))),
-                (
-                    wk::CLASS_TYPES,
-                    Value::Array(vec![Value::ResourceRef(iri("urn:eigenius:lean:LeanExpr"))]),
-                ),
-            ],
-        );
-        builder.add_resource(prop).unwrap();
-        // Holder class for the option_a tests' propositional carriers.
-        // The validator now requires every resource to have at least
-        // one is_a class — this declares a minimal placeholder class
-        // (no required / recommended properties) so the holder
-        // resources can satisfy that without inheriting any
-        // class-typing constraints that would interfere with the test.
-        let holder_class = make_resource(
-            "urn:eigenius:test:PropositionHolder",
-            vec![
-                (
-                    wk::IS_A,
-                    Value::Array(vec![Value::ResourceRef(iri(wk::CLASS))]),
-                ),
-                (wk::SHORT_NAME, Value::String("PropositionHolder".into())),
-                (
-                    wk::DESCRIPTION,
-                    Value::String("test placeholder class".into()),
-                ),
-            ],
-        );
-        builder.add_resource(holder_class).unwrap();
-        Arc::new(builder.build(crate::layer::LayerStorage::in_memory()))
-    }
-
-    /// Option A: `data_type: core:resource` with an `InductiveType`
-    /// `class_types` accepts a single `Value::Json` carrying the
-    /// inductive value, and the validator walks it the same way as
-    /// `data_type: core:inductive`.
-    #[test]
-    fn option_a_resource_with_inductive_class_types_accepts_json() {
-        let layer = build_lean_expr_property_layer(wk::RESOURCE);
-        let mut top = LayerBuilder::new("test_top", Some(layer));
-        let holder = make_resource(
-            "urn:eigenius:test:p_single",
-            vec![
-                (
-                    wk::IS_A,
-                    Value::Array(vec![Value::ResourceRef(iri(
-                        "urn:eigenius:test:PropositionHolder",
-                    ))]),
-                ),
-                (
-                    "urn:eigenius:test:proposition_value",
-                    Value::Json(lambda_x_in_nat()),
-                ),
-            ],
-        );
-        top.add_resource(holder).unwrap();
-        let layer = Arc::new(top.build(crate::layer::LayerStorage::in_memory()));
-
-        let errors = Validator::new(Arc::clone(&layer)).validate();
-        assert!(
-            errors.is_empty(),
-            "`resource` + InductiveType class_types must accept a well-formed Json value; got {errors:?}"
-        );
-    }
-
-    /// Option A: `data_type: core:resource_array` with an
-    /// `InductiveType` `class_types` accepts an `Array` of
-    /// `Value::Json` elements; each element is walked against the
-    /// declared inductive.
-    #[test]
-    fn option_a_resource_array_with_inductive_class_types_accepts_json_array() {
-        let layer = build_lean_expr_property_layer(wk::RESOURCE_ARRAY);
-        let mut top = LayerBuilder::new("test_top", Some(layer));
-        let holder = make_resource(
-            "urn:eigenius:test:p_array",
-            vec![
-                (
-                    wk::IS_A,
-                    Value::Array(vec![Value::ResourceRef(iri(
-                        "urn:eigenius:test:PropositionHolder",
-                    ))]),
-                ),
-                (
-                    "urn:eigenius:test:proposition_value",
-                    Value::Array(vec![
-                        Value::Json(lambda_x_in_nat()),
-                        Value::Json(lambda_x_in_nat()),
-                    ]),
-                ),
-            ],
-        );
-        top.add_resource(holder).unwrap();
-        let layer = Arc::new(top.build(crate::layer::LayerStorage::in_memory()));
-
-        let errors = Validator::new(Arc::clone(&layer)).validate();
-        assert!(
-            errors.is_empty(),
-            "`resource_array` + InductiveType class_types must accept Array<Json>; got {errors:?}"
-        );
-    }
-
-    /// Option A: malformed inductive value in a `resource_array`
-    /// element surfaces as `InductiveValueMismatch` with a structured
-    /// path indicating the bad index — the dispatch in
-    /// `check_class_types` walks each Json element.
-    #[test]
-    fn option_a_resource_array_with_bad_ctor_rejects() {
-        let layer = build_lean_expr_property_layer(wk::RESOURCE_ARRAY);
-        let mut top = LayerBuilder::new("test_top", Some(layer));
-        let bogus = serde_json::json!({"ctor": "DoesNotExist", "args": []});
-        let holder = make_resource(
-            "urn:eigenius:test:p_array_bad",
-            vec![(
-                "urn:eigenius:test:proposition_value",
-                Value::Array(vec![Value::Json(lambda_x_in_nat()), Value::Json(bogus)]),
-            )],
-        );
-        top.add_resource(holder).unwrap();
-        let layer = Arc::new(top.build(crate::layer::LayerStorage::in_memory()));
-
-        let errors = Validator::new(Arc::clone(&layer)).validate();
-        let inductive_errors: Vec<_> = errors
-            .iter()
-            .filter(|e| e.rule == ValidationRule::InductiveValueMismatch)
-            .collect();
-        assert!(
-            !inductive_errors.is_empty(),
-            "bad ctor in array must surface as InductiveValueMismatch; got {errors:?}"
-        );
-        let saw_index_path = inductive_errors.iter().any(|e| e.message.contains("[1]"));
-        assert!(
-            saw_index_path,
-            "error message should reference the failing array index `[1]`; got {inductive_errors:?}"
-        );
-    }
-
-    /// Regression: a `resource_array` property with a Class
-    /// `class_types` still rejects a `Value::Json` element at the
-    /// wire-shape check — Option A only loosens the gate when
-    /// `class_types` resolves to an `InductiveType`.
-    #[test]
-    fn option_a_resource_array_with_class_class_types_rejects_json() {
-        let ctx = crate::bootstrap::bootstrap().expect("bootstrap");
-        let head = Arc::clone(ctx.head());
-
-        let mut builder = LayerBuilder::new("test_class_array", Some(head));
-        // Declare a small Class so class_types resolves.
-        let some_class = make_resource(
-            "urn:eigenius:test:SomeClass",
-            vec![(
-                wk::IS_A,
-                Value::Array(vec![Value::ResourceRef(iri(wk::CLASS))]),
-            )],
-        );
-        let prop = make_resource(
-            "urn:eigenius:test:class_array_prop",
-            vec![
-                (
-                    wk::IS_A,
-                    Value::Array(vec![Value::ResourceRef(iri(wk::PROPERTY))]),
-                ),
-                (wk::SHORT_NAME, Value::String("class_array_prop".into())),
-                (
-                    wk::DATA_TYPE_PROP,
-                    Value::ResourceRef(iri(wk::RESOURCE_ARRAY)),
-                ),
-                (
-                    wk::CLASS_TYPES,
-                    Value::Array(vec![Value::ResourceRef(iri("urn:eigenius:test:SomeClass"))]),
-                ),
-            ],
-        );
-        builder.add_resource(some_class).unwrap();
-        builder.add_resource(prop).unwrap();
-        let lay = Arc::new(builder.build(crate::layer::LayerStorage::in_memory()));
-
-        let mut top = LayerBuilder::new("test_top", Some(lay));
-        let holder = make_resource(
-            "urn:eigenius:test:p_class",
-            vec![(
-                "urn:eigenius:test:class_array_prop",
-                Value::Array(vec![Value::Json(serde_json::json!({"ctor": "Whatever"}))]),
-            )],
-        );
-        top.add_resource(holder).unwrap();
-        let layer = Arc::new(top.build(crate::layer::LayerStorage::in_memory()));
-
-        let errors = Validator::new(Arc::clone(&layer)).validate();
-        let type_mismatches: Vec<_> = errors
-            .iter()
-            .filter(|e| e.rule == ValidationRule::TypeMismatch)
-            .collect();
-        assert!(
-            !type_mismatches.is_empty(),
-            "Class class_types must keep rejecting Json elements at the wire-shape gate; got {errors:?}"
-        );
-    }
-
-    #[test]
-    fn inductive_value_rejects_nested_arg_type_mismatch() {
-        let nat_layer = build_nat_layer();
-        let mut top = LayerBuilder::new("test_top", Some(nat_layer));
-
-        // succ's arg should be a Nat; supply a JSON string.
-        let bad = make_resource(
-            "urn:eigenius:test:bad_nested",
-            vec![(
-                "urn:eigenius:test:nat_value",
-                Value::Json(serde_json::json!({
-                    "ctor": "succ",
-                    "args": ["not_a_nat"]
-                })),
-            )],
-        );
-        top.add_resource(bad).unwrap();
-        let layer = Arc::new(top.build(crate::layer::LayerStorage::in_memory()));
-
-        let errors = Validator::new(Arc::clone(&layer)).validate();
-        let mismatches: Vec<_> = errors
-            .iter()
-            .filter(|e| e.rule == ValidationRule::InductiveValueMismatch)
-            .collect();
-        assert!(
-            !mismatches.is_empty(),
-            "expected an InductiveValueMismatch for nested arg type mismatch; got {errors:?}"
-        );
-        // Path should mention args[0].
-        let path_match = mismatches.iter().any(|e| e.message.contains("args[0]"));
-        assert!(
-            path_match,
-            "error must include structured path `args[0]`: {mismatches:?}"
         );
     }
